reuko-b1-mini/requirements.txt
transformers
datasets
torch
accelerate
pyyaml
tensorboard
